workers/features-worker/requirements.txt
essentia-tensorflow==2.1b6.dev1110
Flask==2.0.3
gunicorn==21.2.0
numpy==1.19.5
python-dotenv==0.20.0
requests==2.27.1